% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohen_d.R
\name{cohen_d}
\alias{cohen_d}
\title{Calculate Cohen's d effect size}
\usage{
cohen_d(group1, group2)
}
\arguments{
\item{group1}{A numeric vector for the first group.}

\item{group2}{A numeric vector for the second group.}
}
\value{
A single numeric value representing Cohen's d.
}
\description{
Calculate Cohen's d effect size
}
\examples{
cohen_d(c(1, 2, 3, 4, 5), c(3, 4, 5, 6, 7))
}
