% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duel_outcome.R
\name{duel_outcome}
\alias{duel_outcome}
\title{Simulate a Yu-Gi-Oh duel outcome}
\usage{
duel_outcome(name1, atk1, name2, atk2)
}
\arguments{
\item{name1}{A character string. Name of the first monster.}

\item{atk1}{Numeric. Attack points of the first monster.}

\item{name2}{A character string. Name of the second monster.}

\item{atk2}{Numeric. Attack points of the second monster.}
}
\value{
Prints the duel outcome to the console.
}
\description{
Simulate a Yu-Gi-Oh duel outcome
}
\examples{
duel_outcome("Blue-Eyes White Dragon", 3000, "Dark Magician", 2500)
}
