% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_score.R
\name{z_score}
\alias{z_score}
\title{Calculate z-scores}
\usage{
z_score(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector of z-scores with the same length as x.
}
\description{
Calculate z-scores
}
\examples{
z_score(c(1, 2, 3, 4, 5))
z_score(c(2, 4, 4, 4, 5, 5, 7, 9))
}
